requirements.txt
# requirements.txt

# Python library requirements
requests==2.26.0
flask
blinker==1.8.1
click==8.1.7
itsdangerous==2.2.0
jinja2==3.1.3
markupsafe==2.1.5
python-dotenv
Werkzeug==3.0.2

psycopg2-binary
pytest
prometheus-flask-exporter
uwsgi
